Remove & ignore macOS `.DS_Store` files
These `.DS_Store` files are used by macOS Desktop Services to store metadata and custom attributes such as folder view options that may change frequently and lead to spurious diffs when versioned in Git repositories.

This removes the files from the repo and tells Git to ignore them in the future.